library refactoring, cmake correction, small fixes
