+++ tests/math_expression_tests/CMakeLists.txt
@@ -8,7 +8,7 @@ if (NOT DEFINED NO_CONAN)
 endif()
 
 add_library(math_expression_test_lib OBJECT 
-    math_expression_t.cpp
+    math_expression_test.cpp
 )
 target_include_directories(math_expression_test_lib PUBLIC
     ${LIBRARIES_DIR}
