+++ tests/math_expression_tests/math_expression_test.cpp
@@ -1,8 +1,7 @@
-#include <math_expression.hpp>
-
 #include <numbers>
 #include <limits>
 
+#include <math_expression/math_expression.hpp>
 #include <boost/ut.hpp>
 
 namespace {
